requirements.txt
fastapi
uvicorn[standard]
python-multipart
moviepy
openai-whisper
srt
ffmpeg-python
